Optimization and some fixes
The wrongGuess(Graphics g) method in hangmanPanel was really bad since
it heavily relied on if statement which were slow and hard to read. The
new code uses a 2D array that stores the coordinates of the lines and
then draws the lines based on the number of incorrect guesses. This new
method is significantly faster and easier to understand.